+++ quickbi-public-20220101/alibabacloud_quickbi_public20220101/models/__init__.py
@@ -98,6 +98,9 @@ from ._data_source_blood_response import DataSourceBloodResponse
 from ._delay_ticket_expire_time_request import DelayTicketExpireTimeRequest
 from ._delay_ticket_expire_time_response_body import DelayTicketExpireTimeResponseBody
 from ._delay_ticket_expire_time_response import DelayTicketExpireTimeResponse
+from ._delete_authorization_by_user_id_request import DeleteAuthorizationByUserIdRequest
+from ._delete_authorization_by_user_id_response_body import DeleteAuthorizationByUserIdResponseBody
+from ._delete_authorization_by_user_id_response import DeleteAuthorizationByUserIdResponse
 from ._delete_data_level_permission_rule_users_request import DeleteDataLevelPermissionRuleUsersRequest
 from ._delete_data_level_permission_rule_users_response_body import DeleteDataLevelPermissionRuleUsersResponseBody
 from ._delete_data_level_permission_rule_users_response import DeleteDataLevelPermissionRuleUsersResponse
@@ -140,6 +143,9 @@ from ._get_user_group_info_response import GetUserGroupInfoResponse
 from ._get_works_embed_list_request import GetWorksEmbedListRequest
 from ._get_works_embed_list_response_body import GetWorksEmbedListResponseBody
 from ._get_works_embed_list_response import GetWorksEmbedListResponse
+from ._ip_white_list_config_request import IpWhiteListConfigRequest
+from ._ip_white_list_config_response_body import IpWhiteListConfigResponseBody
+from ._ip_white_list_config_response import IpWhiteListConfigResponse
 from ._list_acceleration_of_workspace_request import ListAccelerationOfWorkspaceRequest
 from ._list_acceleration_of_workspace_response_body import ListAccelerationOfWorkspaceResponseBody
 from ._list_acceleration_of_workspace_response import ListAccelerationOfWorkspaceResponse
@@ -368,6 +374,9 @@ from ._smartq_authorize_response import SmartqAuthorizeResponse
 from ._smartq_query_ability_request import SmartqQueryAbilityRequest
 from ._smartq_query_ability_response_body import SmartqQueryAbilityResponseBody
 from ._smartq_query_ability_response import SmartqQueryAbilityResponse
+from ._transfer_usergroup_request import TransferUsergroupRequest
+from ._transfer_usergroup_response_body import TransferUsergroupResponseBody
+from ._transfer_usergroup_response import TransferUsergroupResponse
 from ._update_cube_by_sql_request import UpdateCubeBySqlRequest
 from ._update_cube_by_sql_response_body import UpdateCubeBySqlResponseBody
 from ._update_cube_by_sql_response import UpdateCubeBySqlResponse
@@ -417,6 +426,7 @@ from ._get_mail_task_status_response_body import GetMailTaskStatusResponseBodyRe
 from ._get_user_group_info_response_body import GetUserGroupInfoResponseBodyResult
 from ._get_works_embed_list_response_body import GetWorksEmbedListResponseBodyResultData
 from ._get_works_embed_list_response_body import GetWorksEmbedListResponseBodyResult
+from ._ip_white_list_config_response_body import IpWhiteListConfigResponseBodyResult
 from ._list_acceleration_of_workspace_response_body import ListAccelerationOfWorkspaceResponseBodyResultData
 from ._list_acceleration_of_workspace_response_body import ListAccelerationOfWorkspaceResponseBodyResult
 from ._list_api_datasource_response_body import ListApiDatasourceResponseBodyResultData
@@ -630,6 +640,9 @@ __all__ = [
     DelayTicketExpireTimeRequest,
     DelayTicketExpireTimeResponseBody,
     DelayTicketExpireTimeResponse,
+    DeleteAuthorizationByUserIdRequest,
+    DeleteAuthorizationByUserIdResponseBody,
+    DeleteAuthorizationByUserIdResponse,
     DeleteDataLevelPermissionRuleUsersRequest,
     DeleteDataLevelPermissionRuleUsersResponseBody,
     DeleteDataLevelPermissionRuleUsersResponse,
@@ -672,6 +685,9 @@ __all__ = [
     GetWorksEmbedListRequest,
     GetWorksEmbedListResponseBody,
     GetWorksEmbedListResponse,
+    IpWhiteListConfigRequest,
+    IpWhiteListConfigResponseBody,
+    IpWhiteListConfigResponse,
     ListAccelerationOfWorkspaceRequest,
     ListAccelerationOfWorkspaceResponseBody,
     ListAccelerationOfWorkspaceResponse,
@@ -900,6 +916,9 @@ __all__ = [
     SmartqQueryAbilityRequest,
     SmartqQueryAbilityResponseBody,
     SmartqQueryAbilityResponse,
+    TransferUsergroupRequest,
+    TransferUsergroupResponseBody,
+    TransferUsergroupResponse,
     UpdateCubeBySqlRequest,
     UpdateCubeBySqlResponseBody,
     UpdateCubeBySqlResponse,
@@ -949,6 +968,7 @@ __all__ = [
     GetUserGroupInfoResponseBodyResult,
     GetWorksEmbedListResponseBodyResultData,
     GetWorksEmbedListResponseBodyResult,
+    IpWhiteListConfigResponseBodyResult,
     ListAccelerationOfWorkspaceResponseBodyResultData,
     ListAccelerationOfWorkspaceResponseBodyResult,
     ListApiDatasourceResponseBodyResultData,