+++ quickbi-public-20220101/alibabacloud_quickbi_public20220101/models/_create_ticket_request.py
@@ -17,65 +17,46 @@ class CreateTicketRequest(DaraModel):
         watermark_param: str = None,
         works_id: str = None,
     ):
-        # The user\\"s account name.
+        # The account name of the user.
         # 
-        # - If the user is an Alibaba Cloud primary account **wangwu**, the format is **[Primary Account]**, for example, **wangwu**.
+        # - If the user is an Alibaba Cloud account **wangwu**, the format is **[primary account]**, for example, **wangwu**.
+        # - If the user is a Resource Access Management (RAM) users account **zhangsan**@aliyun.cn**, the format is **[primary account:RAM user]**, for example, **wangwu:zhangsan**.
         # 
-        # - If the user is a RAM account **zhangsan**@aliyun.cn\\*\\*, the format is **[Primary Account: Sub-Account]**, for example, **wangwu:zhangsan**.
-        # 
-        # > Only one of UserId and AccountName needs to be filled in. If neither is filled in, it will default to binding the report\\"s Owner, and the report will be accessed with that user\\"s identity. If you need to configure row-level permissions, please refer to [Row-Level Permissions](https://help.aliyun.com/document_detail/322783.html).
+        # > Specify either UserId or AccountName. If neither is specified, the report owner is attached by default, and the report is accessed under that user\\"s identity. To configure row-level permissions for data, see [Row-level permissions](https://help.aliyun.com/document_detail/322783.html).
         self.account_name = account_name
-        # The type of the user\\"s account.
-        # 
+        # The account type of the user. Valid values:
         # - 1: Alibaba Cloud account
-        # 
-        # - 3: Quick BI self-built account
-        # 
+        # - 3: Quick BI custom account
         # - 4: DingTalk
-        # 
-        # - 5: RAM sub-account
-        # 
+        # - 5: RAM user
         # - 9: WeCom
-        # 
-        # - 10: Feishu
-        # 
-        # > If AccountName is not empty, then AccountType must also not be empty.
+        # - 10: Lark
+        # > If AccountName is specified, AccountType must also be specified.
         self.account_type = account_type
-        # Component ID. This is the ID of a component within the above-mentioned dashboard; other types of works do not support this.
-        # Refer to [QueryWorksBloodRelationship](https://next.api.aliyun.com/api/quickbi-public/2022-01-01/QueryWorksBloodRelationship?spm=a2c4g.11186623.0.0.15615d7aWVvWAl\\&params=%7B%7D\\&lang=JAVA\\&tab=DOC\\&sdkStyle=old) for the API to get the component ID.
+        # The component ID. This is the ID of a specific component in the dashboard. Other report types are not supported.
+        # To obtain the component ID, see [QueryWorksBloodRelationship](https://next.api.aliyun.com/api/quickbi-public/2022-01-01/QueryWorksBloodRelationship?spm=a2c4g.11186623.0.0.15615d7aWVvWAl&params={}&lang=JAVA&tab=DOC&sdkStyle=old).
         self.cmpt_id = cmpt_id
-        # Expiration time
-        # 
-        # - Unit: minutes
-        # 
-        # - Default: 240
+        # The expiration time.
+        # - Unit: minutes.
+        # - Default value: 240.
         self.expire_time = expire_time
-        # Global parameters for the report filter conditions.
-        # 
-        # - A string in JsonArray format.
-        # 
-        # > If you need to use global parameter capabilities, please contact the [Quick BI Operations Manager](https://h5-alimebot.dingtalk.com/intl/index.htm?spm=a2c4g.11186623.0.0.3da14f6chrDv9e\\&sourceType=ding_talk\\&from=DEFFB9G5KBByQkwq23wneFIOmaJ).
+        # The global parameter.
         self.global_param = global_param
-        # The number of tickets. Each time a ticket is used, the number of tickets decreases by 1.
-        # 
-        # - Default value: 1
-        # 
-        # - Recommended value: 1
-        # 
-        # - Maximum value: 99999
+        # The number of times the ticket can be used. Each time the ticket is used for access, the count decreases by 1.
+        # - Default value: 1.
+        # - Recommended value: 1.
+        # - Maximum value: 99999.
         self.ticket_num = ticket_num
-        # Quick BI\\"s UserId, which is not your Alibaba Cloud account ID.
-        # You can call the [QueryUserInfoByAccount](https://next.api.aliyun.com/api/quickbi-public/2022-01-01/QueryUserInfoByAccount?spm=a2c4g.11186623.0.0.15615d7aWVvWAl\\&params=%7B%7D\\&tab=DOC\\&sdkStyle=old) interface to obtain the UserId. An example of a UserId is fe67f61a35a94b7da1a34ba174a7\\*\\*\\*\\*.
+        # The Quick BI user ID, not your Alibaba Cloud account ID.
+        # You can call the [QueryUserInfoByAccount](https://next.api.aliyun.com/api/quickbi-public/2022-01-01/QueryUserInfoByAccount?spm=a2c4g.11186623.0.0.15615d7aWVvWAl&params={}&tab=DOC&sdkStyle=old) operation to obtain the user ID. Example: fe67f61a35a94b7da1a34ba174a7****.
         # 
-        # > Only one of UserId and AccountName needs to be filled in. If neither is filled in, it will default to binding the report\\"s Owner, and the report will be accessed with that user\\"s identity. If you need to configure row-level permissions, please refer to [Row-Level Permissions](https://help.aliyun.com/document_detail/322783.html).
+        # > Specify either UserId or AccountName. If neither is specified, the report owner is used by default, and the report is accessed under that user\\"s identity. To configure row-level permissions for data, see [Row-level permissions](https://help.aliyun.com/document_detail/322783.html).
         self.user_id = user_id
-        # Watermark parameters for the report.
-        # 
-        # - Must not exceed 50 characters.
-        # 
-        # - When the report type is a large screen, watermark parameter passing is not supported.
+        # The watermark parameter for the report.
+        # - The value cannot exceed 50 characters.
+        # - Watermark parameters are not supported when the report type is data screen.
         self.watermark_param = watermark_param
-        # The ID of the report to be embedded. Currently supports dashboards, spreadsheets, data screens, self-service data retrieval, ad-hoc analysis, and data entry.
+        # The ID of the report for which embedding is enabled. Dashboards, workbooks, data screens, ad hoc queries, ad hoc analyses, and data entry forms are supported.
         # 
         # This parameter is required.
         self.works_id = works_id