+++ quickbi-public-20220101/alibabacloud_quickbi_public20220101/models/_query_user_by_mobile_account_request.py
@@ -10,8 +10,21 @@ class QueryUserByMobileAccountRequest(DaraModel):
         mobile_type: str = None,
         mobile_user_id: str = None,
     ):
+        # The bound mobile type.
+        # 
+        # - DingTalk: ding
+        # - WeCom: corp_weixin
+        # - Lark: feishu.
+        # 
         # This parameter is required.
         self.mobile_type = mobile_type
+        # The bound mobile user ID.
+        # 
+        # - DingTalk: The unionId of the DingTalk account.
+        # - WeCom: The userId of the WeCom account.
+        # - Lark: The userId of the Lark account.
+        # >Notice: The mobileUserId must be obtained by calling the relevant DingTalk, WeCom, or Lark API operations..
+        # 
         # This parameter is required.
         self.mobile_user_id = mobile_user_id