+++ quickbi-public-20220101/alibabacloud_quickbi_public20220101/models/_transfer_usergroup_response_body.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from darabonba.model import DaraModel
 
-class UpdateCubeBySqlResponseBody(DaraModel):
+class TransferUsergroupResponseBody(DaraModel):
     def __init__(
         self,
         request_id: str = None,
@@ -13,12 +13,9 @@ class UpdateCubeBySqlResponseBody(DaraModel):
     ):
         # The request ID.
         self.request_id = request_id
-        # The execution result of the operation. Valid values:
-        # - true: The request was successful.
-        # - false: The request failed.
+        # Indicates whether the migration was successful.
         self.result = result
         # Indicates whether the request was successful. Valid values:
-        # 
         # - true: The request was successful.
         # - false: The request failed.
         self.success = success